apps/sla-agent-layer/requirements.txt
# SLA-Agent Layer Requirements

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Kafka
kafka-python==2.0.2

# YAML
pyyaml==6.0.1

# OpenTelemetry
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-instrumentation-fastapi>=0.44b0
opentelemetry-exporter-otlp-proto-grpc>=1.24.0
opentelemetry-instrumentation-requests>=0.44b0
opentelemetry-propagator-b3>=1.24.0
# opentelemetry-propagator-tracecontext está incluído no opentelemetry-api

# Métricas Prometheus
prometheus_client>=0.20.0

# Utilitários
python-dateutil==2.8.2
requests==2.31.0
python-multipart==0.0.6
